Implement Vectors of Trust (RFC 8485) for DI auth
With the account manager we have numbered levels (these could be
considered "scalars of trust"), but DI are aligning to an actual
standard.

The `vtr` parameter is a JSON array of acceptable vectors of
trust ("vot"s).  A vot is made up of one or more vector component
values, separated by dots.

The vector component values are:

- `Cl` = username & password
- `Cm` = MFA

So the vot `"Cl"` means "authenticate with a username & password",
whereas the vot `"Cm"` means "authenticate with MFA" (and just MFA -
not all vots are valid).

Dots denote conjunction.  So the vot `"Cl.Cm"` means "authenticate
with username & password & MFA".  This corresponds to our current
`level1`.

Multiple vots in the vtr denote disjunction.  So the vtr `["Cl",
"Cl.Cm"]` means "authenticate with username & password, or with
username & password & MFA".  This corresponds to our current `level0`.

The actual vot uses is returned in the `vot` claim in the ID token.
So we can check that to see if the user authenticated with MFA or not.