Remove SavePages model and db tables
We are reverting [Save a page #69][1] from release_203. The feature was
only ever deployed to integration for testing. We have since decided to
go in a different direction and will replace this with a differet
feature to meet the user need.

This Commit reverts [04729d62e37f2343f81cb9b275732c9b18c8d3a6][2]

[1]: https://github.com/alphagov/account-api/pull/69
[2]: https://github.com/alphagov/account-api/pull/69/commits/04729d62e37f2343f81cb9b275732c9b18c8d3a6