Remove routes as resources for saved_pages
We are reverting [Save a page #69][1] from release_203. The feature was
only ever deployed to integration for testing. We have since decided to
go in a different direction and will replace this with a new
feature to meet the user need.

This Commit reverts [8510a32ef3bb0f3d70e5aaab3d4aeeb57104b53a][2]

[1]: https://github.com/alphagov/account-api/pull/69
[2]: https://github.com/alphagov/account-api/pull/69/commits/8510a32ef3bb0f3d70e5aaab3d4aeeb57104b53a