Add `legacy_sub` field to `OidcUser`
This commit adds the column and starts saving values for new users.
There will be a follow-up migration to copy over values for existing
users, make the column non-null, and add a unique index.

To handle the change in subject identifiers when we switch to DI,
we'll look up users like so:

```ruby
def self.find_or_create_by_sub!(sub:, legacy_sub:)
  transaction do
    user = OidcUser.find_by(sub: sub)
    return user if user

    user = OidcUser.find_by(legacy_sub: legacy_sub)
    if user
      user.update!(sub: sub)
      return user
    end

    OidcUser.create!(sub: sub, legacy_sub: legacy_sub)
  end
end
```

Pre-DI migration, `sub` and `legacy_sub` will always be the same.

Post-DI migration, the only ambiguous case is when decoding a session:
is the sub in the session old or new?  We'll solve that with the card
to add a version field: we can have a new version for
session-with-DI-sub, and then we know if we get an older session we
need to call the DI userinfo endpoint to look up the new one.