Populate `legacy_sub` and add an index
After writing the commit message of b193159, I realised that we don't
actually want this column to be non-nullable.  Because,
post-migration, we will want to create users who have a `sub` but not
a `legacy_sub`.

Usefully, SQL allows multiple NULL values in a nullable field which
has a unique index.  I imagine this is because NULLs are incomparable
in SQL (both `NULL = NULL` and `NULL != NULL` are false).