Reject pre-migration sessions when using Digital Identity
If we see a user with an old session, then we only have an OAuth token
for the account-manager, not Digital Identity.  So we can't query the
Digital Identity UserInfo endpoint to get their new subject
identifier.

In this case, all we can do is log the user out.  We'll have to
migrate old users when they log in.