Migrate users as they authenticate
When we authenticate a user through Digital Identity, check if we have
a record in our database for them.  If we do, all is well.  If not,
we're in one of two cases;

- This user hasn't logged in to GOV.UK at all before
- This user hasn't logged in to GOV.UK since the migration

We can distinguish the two cases by calling the UserInfo endpoint, and
checking the `govuk-account` claim.  If present, this claim holds the
pre-migration subject identifier, which we can use to look up and
migrate the user in our database.