Add session_version to AccountSession
The only thing currently that distinguishes a v1 account session from
prior version is the existance of the session version. So we can safeuly
update all nil versions to at least v1.

When later changes are made we will need to complicate
`self.deserialise` so that it handles additional versions.

These will also need to be added to the KNOWN_VERSIONS array at the top
of the file. Any newly created session with the app will then stamp the
CURRENT_VERSION onto any newly created sesion.

Unknown or mallicious version numbers raise an error, this is handled in
the way that SessionTooOld is from within `self.deserialise` returning
nil.