Create tombstone records for destroyed users
Currently we don't keep any data when a user gets destroyed, we remove
all the records.  This is good for ensuring we don't keep user data
improperly, but bad for auditing (we don't know when a user was
deleted) or for ending sessions (we'll just re-create a blank user
record).

So, create tombstone records when a user is destroyed.  Alternatively,
we could set a `destroyed_at` timestamp on the `OidcUser` record and
wipe the other data, but I think that's a bit more error prone given
we don't currently need to be able to recover destroyed users.