Implement a fake OIDC client for development mode
We can't really use the live Digital Identity service in development
mode:

1. It makes running this app locally depend on a remote service we
don't control, which is not a good idea.

2. We'd have to register a development client with DI and share the
secrets for it amongst all GOV.UK developers.

So, taking a leaf from gds-sso's book, in development mode let's just
sign the user in to the first `OidcUser` in the database.  By making
our `auth_uri` point directly to the callback in frontend, we bypass
the need for a separate OIDC provider, and we can use the `code`
parameter to pass around whether we need MFA (and other similar
information).