Fetch uncached cacheable attributes on sign in
The DI auth tokens are only valid for three minutes and we're
currently not fetching refresh tokens.  So there's little value in
persisting the access token to make a userinfo request later: if we
need to fetch attributes from userinfo, we should do it as soon as the
user authenticates.