Add endpoint to match a user by email address
This will be used to implement the sniffer page in
email-alert-frontend, where we want to examine the email address the
user has just entered and do one of three things:

- Log the user in to email-alert-frontend if they're logged into the
right GOV.UK account; or

- Tell the user to sign in, if a GOV.UK account exists with the email
address they entered; or

- Fall back to the magic link method if there is no GOV.UK account for
that email address.

email-alert-frontend will use this endpoint like so:

    def match_user_by_email(email)
      response = GdsApi.match_user_by_email(
        email: @address,
        govuk_account_session: account_session_header,
      )

      if response["match"]
        :ok
      else
        :must_reauthenticate
      end
    rescue GdsApi::HTTPNotFound
      :no_account
    end

The alternative would be to add an endpoint to find a user by email
address, and then for email-alert-frontend to do something like this:

    def match_user_by_email(email)
      response = GdsApi.account_api.get_attributes(
        attributes: %w[email],
        govuk_account_session: account_session_header,
      )

      return :ok if response["values"]["email"] == email

      GdsApi.account_api.find_user_by_email(email: email)
      :must_reauthenticate
    rescue GdsApi::HTTPNotFound
      :no_account
    end

which would be slower because it makes two API calls rather than one.