Add a unique index on email address
Since we're going to be searching this table by email address, we
should have an index to make it fast.

We can make this a unique index as we will never have two users with
the same email address.  This is because the email address is used as
the username when logging in, and so there cannot be two users with
the same one.