Add rake task to merge pre- & post-migration users
This is to handle the case where a legacy_subject_id isn't being
returned from the userinfo, and so we're trying to create a new
OidcUser record with the same email address as an existing,
pre-migration, one.

We have one such case so far (brought to light by the new unique index
on the email field).  I've asked DI to check the record for that user,
as they should have been migrated across.  But, so that the user can
log in, we need to merge their two records.

The task will only merge users where:

- the post-migration user has no legacy_sub or email.
- the pre-migration user has the same sub and legacy_sub.
- both users were created on the right sides of the migration.