Make check-email-subscription endpoint also take a base_path
If a base_path is given, we match against the URL field of the
subscriber list.  This is so we can find a subscription for a specific
page, without needing the frontend app to query email-alert-api for
the slug first.

I've also added tests for the case where a user has active
subscriptions which don't match, as that wasn't covered previously.