Omit details of root cause when reporting a sensitive exception
Throwing a `CapturedSensitiveException` inside a `rescue` embeds the
details of the causing exception inside it.  This is usually a useful
feature, letting you drill down into the root cause of a problem.

But when our goal is to hide exceptions which contain user data, it's
not so helpful.

This commit adds a handler for `CapturedSensitiveException`s which
just reports a string (and the ID of the newly-created
`SensitiveException` record) to Sentry.  Since we're not reporting an
actual exception object, there's no stack trace or any other details
being sent to Sentry.