Move internal & personalisation controllers into separate modules
Previously, we had the `ApplicationController` set up to handle
internal API endpoints, by rendering `ApiError` exceptions and by
validating gds-sso users.

Our personalisation endpoint needed to override some of this
behaviour.  As we add more personalisation endpoints, they will need
to do the same.  Having a bunch of classes which override the same
functionality of their superclass smells like the superclass is the
wrong abstraction.

So, this commit introduces new superclasses for internal and
personalisation controllers, with the `ApplicationController` only
containing truly common functionality.

This commit does NOT fix a 401 error from the check-email-subscription
endpoint being returned as:

    {"message":"No bearer token was provided"}

I think that that's because gds-sso pulls in warden, and warden (by
default!) intercepts *all* 401 responses, even such responses from
controllers which don't use warden.  Incidentally, this means that all
our 401-returning internal endpoints are also returning that unhelpful
message.  I think our best option is probably to just not return 401
responses at all, and to treat that as the gds-sso error code.  Not
great, but changing this behaviour of gds-sso feels like it may break
things elsewhere in the GOV.UK stack.