Pass location to email subscription button
It's possible to have two of these buttons on a page - one before and
one after the content. We'd like to be able to track clicks on these
separately in Google Analytics, so we need some way of differentiating
them.

Accept an optional  `button_location` parameter to the check email
subscriptions route and pass it on to the component so the component
can handle adding the required attributes for click tracking.

Also return the `button_location` in the JSON response to aid debugging.