Use `extra`, not `tags`, for captured sensitive exception IDs
`tags` is a way of filtering events.  When looking at a group of
events, you can click the "tags" panel, and then click a tag value to
filter to only those events with the same tag.

`extra` is a way of adding additional information to a single event.
When looking at a group of events, each has an "Additional Data"
section (beneath the "Operating System" and "Runtime" sections).  No
changing of panels needed, it's right there.

So change to `extra` here as it's more correct (and more convenient)
for what we're trying to do: attach unique data to each event.