Remove use of refresh tokens entirely
Our pre-DI auth prototype service issued refresh tokens.  We used
these to fetch the UserInfo response when a user tried to use an
attribute which we don't have cached for them.

DI do not do refresh tokens, so we made the minimal change needed to
our code: we added support for there *not* being a refresh
token (rather than assuming one always exists) and also made the
attribute-caching happen as soon as the access token is fetched, not
when the attribute is needed.  But we left the ability to use refresh
tokens in place.

DI have spotted us using invalid access tokens on occasion, and have
suggested that we might be re-using tokens.  This is kind of weird as
we now only use the access token to fetch the userinfo in the auth
callback: there are two places where we do that, but we share the
userinfo response if it gets fetched in the first place.  So we
shouldn't try to use the same access token twice.

I suspect that the change we made to support DI was not quite right.

If a userinfo request fails due to a 4xx error (which should ideally
not happen anyway), we'll call `OidcClient#refresh_client_tokens` and
retry the request.  We don't have anything to skip that if the refresh
token is missing, so my hunch is that that function just returns the
access token the client already has, and so one token is getting used
a second time.

Since we don't need refresh tokens in production, we can simplify the
OidcClient by stripping them out entirely.  If access tokens are still
being reused, this wasn't the real problem and we'll need to do some
more digging.  But even if that is the case, this change is positive.