Remove references to transition checker
This commit removes all the code and tests referencing the
transition_checker_state attribute, and relevant examples from the
docs.

In a follow-up commit deployed after this one, we should add a
database migration to drop these from the database:

- `oidc_users.has_received_transition_checker_onboarding_email`
- `oidc_users.transition_checker_state`
- `unmigrated_oidc_users`

One downside of this change is that we no longer have an attribute
which stores arbitrary JSON, so the "correctly round-trips local
attributes" test is now much weaker: it's only testing that we handle
booleans properly.  I considered adding a test attribute, but that
felt like a step too far.