Remove old transition checker state
This commit removes the transition checker state from the `oidc_users`
table, and drops the `unmigrated_oidc_users` table entirely.  This
migration is not reversible.

It's worth noting that we've not had any tickets which needed us to
check whether a user existed but wasn't migrated since the original
flurry of migration tickets.  So dropping the entire
`unmigrated_oidc_users` table, rather than just the column for saved
transition checker results, seems sensible.  If we do need this
information, we can always consult the backup we took of the
account-manager database before shutting that app down.