Add extra_info to sensitve exception
Whilst trying to debug we need more information about a subset of
errors. Here we add a next text string (intended for string encoded
JSON), and then supply it with a response code, response body and access
token info for a certain category of error.

We will want to remove this once we're done.