Log tokens response for failed userinfo calls
This is a temporary measure to help debugging and has no effect on
successful logins.

Store the processed response from the tokens endpoint in an instance
variable. Then, if the following call to userinfo fails, include the
tokens response in the `SensitiveException` information. This will
allow us to ensure that the tokens match what we send to the OIDC
provider and also check the time between calls (as the access token has
a low TTL).