Add Logout Token Validation
We're implimenting OIDC backchannel logout. The spec shows we have to do
quite a bit we'd have to do to validate and verify a LogoutToken before
we take an action.

This is a mix of comparing data in the token to Digital Identity's
well-known data. Data we logged at registration (our client_id) and then
a bunch of formatting and logical consistency checks.

It makes sense to make this it's own service object, so that's what
we've done here with releavant tests.

There's more going on in an initaliser here than I'd usually want in a
service object. Plus we've included a gem attr_required which is non
standard. This is somewhat to make the token class confirm to what we
find in the OpenIDConnect IDToken class.

The spec calls out that the validation rules for these is similar, so
I've based this off [the example provided][5] by [the openid_connect
gem][2].  That was partly to get my head around how a token functions
here, and given we already use [a class method in our OIDC
validation][4].

I'm somewhat hopeful we could eventually mainstream this back up to the
gem and have this provided along with the rest of our OIDC token
validation.  I've opend an issue [to discuss here][3] and will put in a
PR if they're open to it.

[1]: https://openid.net/specs/openid-connect-backchannel-1_0.html#Validation
[2]: https://github.com/nov/openid_connect
[3]: https://github.com/nov/openid_connect/issues/67
[4]: https://github.com/alphagov/account-api/blob/f73ddbd4b8f36b6fe63f07993d5e1c8711652246/app/lib/oidc_client.rb#L79
[5]: https://github.com/nov/openid_connect/blob/master/lib/openid_connect/response_object/id_token.rb