Extend OidcClient to add a logout_token endpoint
This will recieve an encoded JWT:
- Verify the signature (or throw an error)
- Then ask the LogoutToken class to validate the payload

Assuming all is well there it'll return a hash with a request_time, the
token as a JWT and the token itself. Mostly trying to keep pairty with
the id_token inclusion in the tokens! method. But I can see how those'd
be useful for debugging too.

Testing this is a bit of a pain. Discovery reaches out to the
.well-known endpoints of the Identity Provider. So I've stubbed those
out. I also skipped the JWS payloads and just used a secret that's a
string. The JSON::JWT code seems smart enough to work out what to do
with all of this.

The important bit is that we're following the backchannel logout spec by
checking signatures for the logout token in the same way we're doing for
the ID token.