Remove cookies and feedback consent attributes
We've removed the first time sign in page which asked users to consent
to cookies and sending them emails for feedback [[1]] and the form to
see and manage these attributes in the account home area [[2]].

Now we can remove them from the API and drop the database columns so
we're not hoarding personal data.

These two attributes are the only two `local` attributes, and are used
throughout for testing the logic for local vs. cached attributes.
Instead of removing all that logic as well, this commit sets up a new
dummy attribute `local_attribute` that can be used for testing while
keeping the logic in place. A later PR will clean up the different types
of attributes.

[1]: https://github.com/alphagov/frontend/pull/3234
[2]: https://github.com/alphagov/frontend/pull/3225