Add a `local_attribute` attribute for testing
We are planning to remove the cookie and feedback consent attributes
as these are no longer going to be stored in the account API. However,
these are the last two local attributes and are used throughout the test
suites to test the logic for local attributes vs. cached attributes.

Removing them would therefore cause all the unit tests to fail. Rather
than deleting all the logic differentiating between attribute types at
this stage (which will take a while), we are adding a new local
attribute which is only used for testing.

We will then update the pact tests to use this new attribute, then
remove the cookie and feedback consent attributes and drop the data.

We'll then be in a place where we can clean up the tech debt introduced
by this commit and remove all the logic for local attributes.