Allow multiple Tombstone records for the same sub
When this app was originally created it was *the* canonical place to
store account data. The Tombstone records were to allow us to count
deleted accounts and enforce uniqueness.

Now however, that's not the case. The source for account data is Auth's
OICD provider and Tombstones are less helpful to us now this app is
only part of a relying party's account storage.

Accounts on GOV.UK are only used for email notifications, so we plan to
use email alert API's logic to delete accounts that have no subscriptions.

This means it will be possible for someone to create an account, sign up
for notifications, delete that subscription and then have that account
deleted aoutmatically. We need to make sure that if that person comes
back to GOV.UK and signs in, we can create a new `OidcUser` for them
that behaves exactly as expected.

In order for that to happen, we have to be able to create multiple
Tombstones for the same sub.