Delete old Tombstone records
In [[1]] we allowed there to be multiple Tombstones with the same sub
as this app is no longer the canonical place for account creation and
deletion data.

Since we're no longer checking for duplicates, we can also start
deleting old records as they're now only really useful for debugging.

We went with 30 days as that matches the time needed for an account to
be marked 'inactive' by email alert API.

We do lose the ability to count the number of accounts that has ever
existed by adding this worker but, as mentioned before, this app isn't
the right place to count this anyway so it's not a real loss.

[1]: https://github.com/alphagov/account-api/pull/421