GUA-22 Invalidate LogoutNotice on successful /callback
The callback endpoint is used at the end of a successful journey with
the OP (login / create). In either circumstance the user has indicated a
desire to be "logged in". This invalidates and prior request to "be
logged out" so we can safely remove a notice to "globally log the user
out".

To make it to the point of invalidation, the callback must be posted
correctly + have created an AccountSession object without raising an
OidcClient::OAuthFailure.

Invalidating the notice will be teh last thing we do before returning
JSON