GUA-22 Invalidate LogoutNotice on deletion of OIDC user
If we recieve a request from an OP or from the email-alert-api to delete
an OIDC user then we should also scrub any outstanding LogoutNotices.

Users that don't exist cannot log in.

Users that later are resurrected by Oidc Magic will either have a new
Sub if the delete was from the OP, or will have logged in via the OP and
so will have again showed a new intention to "be signed in" that
invalidates the LogoutNotice