Remove LogoutNotice if we create a new user with a repeated sub
In the case that a user has unsubscribed from their last email
notification we will tidy up their account after a period of time (to
prevent us holding onto details we have no user for!).

In this case we'll delete their OidcUser record. The user will still
have the same sub over on the OP (digital identity). If they later log
in with digital identity and look to create a new subscription we'll
need to recreate their account.

Whilst we've set LogoutNotices to be removed upon delete of an account,
for saftey it seems prudent to take any indiciation that we need to
create an OidcUser (creating an account) as an indication that the user
will want to log in and use it.

Those circumstances invalidate any previous desire to "be logged out"
for a user with the same sub.

So we'll remove the logout notice in that case