Remove attribute types and local attribute logic
In #418 we removed the last two local attributes, cookie and feedback
consent. In that PR we replaced them with a dummy local attribute that
allowed all the logic to stay in place temporarily.

This commit cleans up after that PR by removing attribute types and the
logic associated with what used to be `local` attributes. We only have
what used to be `cached` attributes left, which are those provided by
the IDP and aren't allowed to be updated by internal GOV.UK apps.

We might want to add other attributes in the future and we don't want
to break existing behavoiur when trying to update these attributes
(which will return a 403), so this commit removes the logic for
updating attributes and the associated tests but leaves the `update`
controller method in place. This means the change to gds-api-adaptors
doesn't need to be a breaking change where we remove a method and have
to audit all other apps to check if it's used.