Use FileStore for the ActiveSupport cache.
Use
[FileStore](https://guides.rubyonrails.org/caching_with_rails.html#activesupport-cache-filestore)
instead of MemCacheStore for caching the OIDC discovery response YAML.

This cache isn't used for anything else, and if its use expands in the
future then it's easy to change it again later.

This saves Replatforming from configuring and running an entire memcache
just to store a kilobyte of YAML that doesn't change 😅

In the EC2 deployment this cache is already per-instance, so there's no
significant regression in moving to a per-pod cache.

Tested: runs without errors on k8s integration.

Co-authored-by: Marc Pomfret <marc.pomfret@digital.cabinet-office.gov.uk>
Co-authored-by: Martyn Pratt <martyn.pratt@digital.cabinet-office.gov.uk>