GUA-838: Add new link to /sign-in page
In the past quarter the top 4 and 5 on page searches were for 'visa' or
'visa application' and users were going to the visa application guidance
page. We have added a link to the guidance page on 'How to sign in to
continue your visa application'. Some supporting data suggests that
people were struggling to sign in to complete an unfinished visa
application